refactor: 이용약관 등 static 폴더로 이동
