+++ src/components/static/PrivacyPolicy.js
