+++ src/components/static/Terms.js
